Changed field name for soft cascades to "softCascades"

+++ src/Mbcraft/Laravel/Database/SoftDeletesCascade.php
@@ -8,13 +8,15 @@ trait SoftDeletesCascade {
         
         if (method_exists($this, "runSoftDelete")) {
             $softCascades = $this->softCascades;
-            
-            foreach ($softCascades as $access_method) {
-                $cascade_objects = $this->{$access_method}()->get()->toBase()->all();
-                foreach ($cascade_objects as $objs) {
-                    $objs->delete();
+            if ($softCascades!=null) {
+                foreach ($softCascades as $access_method) {
+                    $cascade_objects = $this->{$access_method}()->get()->toBase()->all();
+                    foreach ($cascade_objects as $objs) {
+                        $objs->delete();
+                    }
                 }
-            }
+            } else
+                \Log::debug("softCascades are null for Player!!");
         }
         
         parent::delete();
